+++ client/qml/Timeline.qml
@@ -226,22 +226,6 @@ Rectangle {
         }
     }
 
-    DropArea {
-        anchors.fill: parent
-        onEntered: if (!room) drag.accepted = false
-        onDropped: {
-            if (drop.hasUrls) {
-                controller.fileDrop(drop.urls)
-                drop.acceptProposedAction()
-            } else if (drop.hasHtml) {
-                controller.htmlDrop(drop.html)
-                drop.acceptProposedAction()
-            } else if (drop.hasText) {
-                controller.textDrop(drop.text)
-                drop.acceptProposedAction()
-            }
-        }
-    }
     ScrollView {
         id: chatScrollView
         anchors.top: roomHeader.bottom
